+++ src/main/java/com/aleksander/restaurant/reservation/repository/ReservationRepository.java
@@ -6,10 +6,11 @@ import com.aleksander.restaurant.reservation.model.ReservationStatus;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository extends JpaRepository<Reservation, Long>, JpaSpecificationExecutor<Reservation> {
     List<Reservation> findByTableId(Long tableId);
 
     List<Reservation> findByTableIdAndStatus(Long tableId, ReservationStatus status);